requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
tensorflow==2.17.0
tensorflow-intel==2.17.0
numpy==1.26.4
pandas==2.2.2
h5py==3.11.0
Pillow==10.4.0
openai==1.43.0
geopy==2.4.1
pydantic==2.8.2
requests==2.32.3
tensorboard==2.17.1
